Refactor the code structure of aws directory
