+++ aws/query_test.go
